controller tests working and successful

+++ src/main/java/com/revature/gambit/skill/services/SkillService.java
@@ -2,9 +2,11 @@ package com.revature.gambit.skill.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import java.util.List;
 
 import com.revature.gambit.skill.beans.Skill;
 import com.revature.gambit.skill.repo.SkillRepository;
@@ -75,6 +77,13 @@ public class SkillService implements ISkillService {
 	public Skill saveSkill(Skill skill) {
 		return skillRepository.saveAndFlush(skill);
 	}
+	
+	@Override
+	@Transactional
+	public Skill deleteSoftly(String name) {
+		Skill skill = findByName(name);
+		return skillRepository.saveAndFlush(skill);
+	}
 
 	/**
 	 * Deletes a skill based on its name.
@@ -83,8 +92,13 @@ public class SkillService implements ISkillService {
 	 *            Name of the skill to delete.
 	 */
 	@Override
-	public void deleteSkillViaName(String name) {
-		skillRepository.delete(findByName(name));
+	@Transactional
+	public boolean deleteSkillViaName(String name) {
+		if (findByName(name) instanceof Skill) {
+			skillRepository.delete(findByName(name));
+			return true;
+		}
+		return false;
 	}
 
 }
