+++ src/main/java/com/revature/gambit/skill/services/SkillTypeService.java
@@ -81,8 +81,12 @@ public class SkillTypeService implements ISkillTypeService {
 	 *            Name of the skill type to delete.
 	 */
 	@Transactional
-	public void deleteBySkillTypeName(String name) {
-		this.skillTypeRepository.deleteBySkillTypeName(name);
+	public boolean deleteBySkillTypeName(String name) {
+		if (findBySkillTypeName(name) instanceof SkillType) {
+			this.skillTypeRepository.deleteBySkillTypeName(name);
+			return true;
+		}
+		return false;
 	}
 
 }
